Kommit

+++ MemberGallery/MemberGallery/Model/Category.cs
@@ -8,11 +8,10 @@ namespace MemberGallery.Model
 {
     public class Category
     {
-        public short CategoryID { get; set; }
+        public int CategoryID { get; set; }
 
         [Required(ErrorMessage = "Fält Kategori får ej lämnas tomt.")]
         [StringLength(20, ErrorMessage = "Kategori kan max vara 20 tecken långt.")]
         public string CategoryProp { get; set; }
-
     }
 }
